db-sync-local/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asn1crypto==1.5.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloud-sql-python-connector[pg8000]==1.15.0
cryptography==44.0.0
dnspython==2.7.0
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
multidict==6.1.0
numpy==2.2.1
pandas==2.2.3
pg8000==1.31.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
requests==2.32.3
rsa==4.9
schedule==1.2.2
scramp==1.4.5
six==1.17.0
sqlalchemy==2.0.36
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
